# openapi.yaml
openapi: 3.0.3
info:
  title: SiteNinja Backend API
  description: |
    Multi-tenant website builder backend built with Next.js 15, TypeScript, Prisma, and PostgreSQL.

    ## Authentication

    This API uses two authentication methods:

    1. **JWT Bearer Token** (Primary)
       - Include in header: `Authorization: Bearer <token>`
       - Get token from `/api/auth/callback/credentials` endpoint

    2. **API Key** (Server-to-server)
       - Include in header: `X-API-Key: <api_key>`
       - Create keys via `/api/tenants/{tenantId}/api-keys`

    ## Rate Limiting

    - Anonymous: 100 requests/hour
    - Authenticated: 1,000 requests/hour
    - API Keys: 5,000 requests/hour

    ## Test Credentials

    After running `npm run db:seed`, you can use:
    - Super Admin: `admin@siteninja.com` / `Password123!`
    - Restaurant Owner: `marco@bellaitalia.com` / `Password123!`
    - Tech Owner: `sarah@techflow.io` / `Password123!`

  version: 2.0.0
  contact:
    name: SiteNinja Support
    email: support@siteninja.com
servers:
  - url: http://localhost:3021
    description: Local development server
  - url: https://api.siteninja.com
    description: Production server
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []
paths:
...
  /api/ai/enhance:
    post:
      tags: [AI Features]
      summary: Enhance content with AI
      description: Use AI to improve content clarity, engagement, and readability
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - content
                - tenantId
              properties:
                content:
                  type: string
                  minLength: 10
                  maxLength: 5000
                  description: The content to enhance
                  example: "We serve delicious pizza and pasta at our restaurant."
                tenantId:
                  type: string
                  format: uuid
                  description: Tenant ID for authorization
                tone:
                  type: string
                  enum: [professional, casual, friendly, formal, creative]
                  default: professional
                  description: Desired tone for the enhanced content
                length:
                  type: string
                  enum: [shorter, longer, similar]
                  default: similar
                  description: Target length compared to original
                focus:
                  type: string
                  maxLength: 200
                  description: Additional focus areas for enhancement
                  example: "make it more engaging and descriptive"
                model:
                  type: string
                  description: Override default AI model
                  example: "gpt-4o"
      responses:
        '200':
          description: Content enhanced successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          original:
                            type: string
                          enhanced:
                            type: string
                          metadata:
                            type: object
                            properties:
                              tone:
                                type: string
                              improvements:
                                type: array
                                items:
                                  type: string
                              wordCount:
                                type: number
                              model:
                                type: string
        '401':
          description: Unauthorized
        '429':
          description: Rate limit exceeded
  /api/ai/seo:
    post:
      tags: [AI Features]
      summary: Generate SEO metadata
      description: Generate optimized SEO metadata using AI
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - content
                - currentTitle
                - tenantId
              properties:
                content:
                  type: string
                  minLength: 50
                  maxLength: 5000
                  description: Page content for SEO analysis
                currentTitle:
                  type: string
                  minLength: 1
                  maxLength: 200
                  description: Current page title
                tenantId:
                  type: string
                  format: uuid
                targetKeywords:
                  type: array
                  items:
                    type: string
                  description: Target keywords for SEO
                businessType:
                  type: string
                  maxLength: 100
                  description: Type of business
                model:
                  type: string
                  description: Override default AI model
      responses:
        '200':
          description: SEO suggestions generated
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          current:
                            type: object
                            properties:
                              title:
                                type: string
                          suggestions:
                            type: object
                            properties:
                              metaTitle:
                                type: string
                                maxLength: 60
                              metaDescription:
                                type: string
                                maxLength: 160
                              keywords:
                                type: array
                                items:
                                  type: string
                              improvements:
                                type: array
                                items:
                                  type: string
  /api/ai/ideas:
    post:
      tags: [AI Features]
      summary: Generate content ideas
      description: Generate creative content ideas using AI
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - tenantId
              properties:
                tenantId:
                  type: string
                  format: uuid
                count:
                  type: integer
                  minimum: 1
                  maximum: 20
                  default: 5
                  description: Number of ideas to generate
                businessType:
                  type: string
                  maxLength: 100
                  description: Type of business
                topic:
                  type: string
                  maxLength: 200
                  description: Specific topic focus
                model:
                  type: string
                  description: Override default AI model
      responses:
        '200':
          description: Content ideas generated
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          ideas:
                            type: array
                            items:
                              type: object
                              properties:
                                title:
                                  type: string
                                description:
                                  type: string
                                category:
                                  type: string
  /api/ai/rewrite:
    post:
      tags: [AI Features]
      summary: Rewrite content with custom instructions
      description: Rewrite content according to specific instructions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - content
                - instructions
                - tenantId
              properties:
                content:
                  type: string
                  minLength: 10
                  maxLength: 5000
                  description: Content to rewrite
                instructions:
                  type: string
                  minLength: 5
                  maxLength: 500
                  description: Custom rewriting instructions
                  example: "Make it sound more gourmet and upscale"
                tenantId:
                  type: string
                  format: uuid
                model:
                  type: string
                  description: Override default AI model
      responses:
        '200':
          description: Content rewritten
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          original:
                            type: string
                          rewritten:
                            type: string
                          metadata:
                            type: object
                            properties:
                              instructions:
                                type: string
                              model:
                                type: string
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from login endpoint

    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for server-to-server authentication
...
  schemas:
...
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
        message:
          type: string
        meta:
          type: object
...